Addressing comments from Kirsten
